refactor: model_utils into shared utils folder with updates to better structure additional utils such as data utils
