+++ pytorch_lattice/models/calibrated_linear.py
@@ -5,13 +5,13 @@ import torch
 
 from ..constrained_module import ConstrainedModule
 from ..layers import Linear
-from .features import CategoricalFeature, NumericalFeature
-from .model_utils import (
+from ..utils.models import (
     calibrate_and_stack,
     initialize_feature_calibrators,
     initialize_monotonicities,
     initialize_output_calibrator,
 )
+from .features import CategoricalFeature, NumericalFeature
 
 
 class CalibratedLinear(ConstrainedModule):
