+++ pytorch_lattice/utils/data.py
@@ -1,0 +1,1 @@
+"""Utility functions and classes for handling data."""
