--- tests/dummy_test.py
@@ -1,6 +1,0 @@
-"""Dummy test for initial structure to pass automatic testing on push."""
-
-
-def test_dummy():
-    """Dummy test."""
-    pass
